# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SeatLock",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/v1/boarding/validate": {
            "post": {
                "operationId": "v1.boarding.validate",
                "summary": "Valida o QR Code no embarque e marca o bilhete como usado (uso \u00fanico).\nRestrito a operadores",
                "tags": [
                    "Boarding"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "maxLength": 1000
                                    },
                                    "trip_id": {
                                        "type": [
                                            "string",
                                            "null"
                                        ]
                                    }
                                },
                                "required": [
                                    "code"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "valid": {
                                                    "type": "boolean"
                                                },
                                                "ticket_id": {
                                                    "type": "string"
                                                },
                                                "used_at": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "passenger": {
                                                    "type": "string"
                                                },
                                                "seat": {
                                                    "type": "string"
                                                },
                                                "trip_id": {
                                                    "type": "string"
                                                },
                                                "leg": {
                                                    "type": "object",
                                                    "properties": {
                                                        "origin": {
                                                            "type": "integer"
                                                        },
                                                        "destination": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "required": [
                                                        "origin",
                                                        "destination"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "valid",
                                                "ticket_id",
                                                "used_at",
                                                "passenger",
                                                "seat",
                                                "trip_id",
                                                "leg"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "An error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Error overview.",
                                            "examples": [
                                                "Apenas operadores validam embarque."
                                            ]
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "ProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "format": "uri",
                        "description": "URI que identifica o tipo do problema."
                    },
                    "title": {
                        "type": "string",
                        "description": "Resumo leg\u00edvel do problema."
                    },
                    "status": {
                        "type": "integer",
                        "description": "Status HTTP."
                    },
                    "detail": {
                        "type": "string",
                        "description": "Explica\u00e7\u00e3o desta ocorr\u00eancia."
                    },
                    "instance": {
                        "type": "string",
                        "description": "URI da requisi\u00e7\u00e3o que falhou."
                    },
                    "errors": {
                        "type": "object",
                        "description": "Erros de valida\u00e7\u00e3o por campo (422).",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "conflicting_seat_ids": {
                        "type": "array",
                        "description": "Assentos indispon\u00edveis (409 seat-unavailable).",
                        "items": {
                            "type": "string"
                        }
                    },
                    "alternatives": {
                        "type": "array",
                        "description": "Sugest\u00f5es de assentos livres parecidos (409 seat-unavailable).",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "number": {
                                    "type": "string"
                                },
                                "type": {
                                    "type": "string"
                                },
                                "deck": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                },
                "required": [
                    "type",
                    "title",
                    "status"
                ],
                "title": "ProblemDetails"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/ProblemDetails"
                        }
                    }
                }
            },
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "$ref": "#/components/schemas/ProblemDetails"
                        }
                    }
                }
            }
        }
    }
}